Restricciones para Servicio de API

+++ src/main/java/com/grupoh/springbootmvc/Service/PaginaService.java
@@ -6,8 +6,8 @@ import com.grupoh.springbootmvc.Repository.PaginaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import javax.persistence.TypedQuery;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -20,13 +20,23 @@ public class PaginaService {
     }
 
 
-    public Pagina getPagina(String titulo) {
-        return paginaRepository.findPaginaByTitulo(titulo);
+    public Pagina getPaginaFromTitulo(String titulo) {
+        Pagina paginaEnc = paginaRepository.findPaginaByTitulo(titulo);
+        if (paginaEnc != null) {
+            return paginaEnc;
+        } else throw new NoSuchElementException(
+                "No existe una pagina con titulo " + titulo + " en la BD"
+        );
     }
 
     // Method to delete all paginas into the DB
     public void deleteAllPaginas() {
-        paginaRepository.deleteAll();
+        if (!paginaRepository.findAll().isEmpty()) {
+            paginaRepository.deleteAll();
+        } else throw new NoSuchElementException(
+                "No se encuentran paginas en la BD"
+        );
+
     }
 
     // Method to save all paginas into the DB
@@ -36,22 +46,40 @@ public class PaginaService {
 
     // Method to get all the paginas in to the DB
     public List<Pagina> getAllPaginas() {
-        return paginaRepository.findAll();
+        List<Pagina> paginas = paginaRepository.findAll();
+        if (!paginas.isEmpty()) {
+            return paginas;
+        } else throw new NoSuchElementException(
+                "No se encuentran páginas en la BD"
+        );
     }
 
     public Optional<Pagina> getPaginaFromId(Long paginaId) {
-        return paginaRepository.findById(paginaId);
+        if (paginaRepository.existsById(paginaId)) {
+            return paginaRepository.findById(paginaId);
+        } else throw new NoSuchElementException(
+                "Pagina con id " + paginaId + " no existe en la BD"
+        );
+
     }
 
     public void addPagina(Pagina pagina) {
-        paginaRepository.save(pagina);
+        Long idPag = pagina.getId();
+        if (idPag != null) {
+            if (!paginaRepository.existsById(idPag)){
+                paginaRepository.save(pagina);
+            } else throw new IllegalStateException(
+                    "La pagina con id " + idPag + " ya existe. Inserte una pagina sin id para que se asigne automaticamente"
+            );
+        } else paginaRepository.save(pagina);
         System.out.println(pagina);
+
     }
 
     public void deletePaginaById(Long paginaId) {
         if (paginaRepository.existsById(paginaId)) {
             paginaRepository.deleteById(paginaId);
-        } else throw new IllegalStateException(
+        } else throw new NoSuchElementException(
                 "Pagina con id " + paginaId + " no existe"
         );
     }
@@ -61,12 +89,17 @@ public class PaginaService {
             pagina.setId(paginaId);
             paginaRepository.save(pagina);
             System.out.println(pagina);
-        } else throw new IllegalStateException(
+        } else throw new NoSuchElementException(
                 "Compania con id " + paginaId + " no existe"
         );
     }
 
     public List<Pagina> getUserPages( Usuario u ) {
-        return paginaRepository.getUserPages(u.getId());
+        List<Pagina> paginasUser = paginaRepository.getUserPages(u.getId());
+        if (!paginasUser.isEmpty()) {
+            return paginasUser;
+        } else throw new NoSuchElementException(
+                "El usuario con id " + u.getId() + " no tiene paginas asignadas"
+        );
     }
 }
